ai_backend/requirements.txt
# FastAPI core
fastapi==0.115.0
uvicorn[standard]==0.30.6

# AI/ML frameworks (PyTorch will be installed separately for CPU-only)
ultralytics==8.3.0
opencv-python==4.10.0.84

# Image processing
pillow==10.4.0
numpy==1.26.4

# File upload support
python-multipart==0.0.9

# Additional utilities
python-dotenv==1.0.1
